Mapping Repo and Service impl

+++ SimplyFly/src/main/java/com/hexaware/simplyfly/entity/Booking.java
@@ -22,7 +22,7 @@ public class Booking {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long bookingId;
+    private int bookingId;
 
     private LocalDateTime bookingDate;
