+++ SimplyFly/src/main/java/com/hexaware/simplyfly/repository/RouteRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.hexaware.simplyfly.entity.Route;
 
 @Repository
-public interface RouteRepository extends JpaRepository<Route, Long> {
+public interface RouteRepository extends JpaRepository<Route, Integer> {
     List<Route> findByOriginAndDestination(String origin, String destination);
 }
