+++ SimplyFly/src/main/java/com/hexaware/simplyfly/repository/SeatRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import com.hexaware.simplyfly.entity.Seat;
 
 @Repository
-public interface SeatRepository extends JpaRepository<Seat, Long> {
-	List<Seat> findByFlightFlightId(Long flightId);
+public interface SeatRepository extends JpaRepository<Seat, Integer> {
+	List<Seat> findByFlightFlightId(int flightId);
 
 	List<Seat> findByIsBooked(boolean isBooked);
 }
