+++ SimplyFly/src/main/java/com/hexaware/simplyfly/service/FlightService.java
@@ -5,11 +5,11 @@ import java.util.List;
 import com.hexaware.simplyfly.entity.Flight;
 
 public interface FlightService {
-	Flight getFlightById(Long id);
+	Flight getFlightById(int id);
 
 	List<Flight> searchFlights(String origin, String destination);
 
-	List<Flight> getFlightsByOwner(Long ownerId);
+	List<Flight> getFlightsByOwner(int ownerId);
 
 	Flight addFlight(Flight flight);
 }
