+++ SimplyFly/src/main/java/com/hexaware/simplyfly/service/PaymentService.java
@@ -5,11 +5,11 @@ import java.util.List;
 import com.hexaware.simplyfly.entity.Payment;
 
 public interface PaymentService {
-	Payment getPaymentById(Long id);
+	Payment getPaymentById(int id);
 
 	List<Payment> getPaymentsByStatus(String status);
 
-	Payment getPaymentByBooking(Long bookingId);
+	Payment getPaymentByBooking(int bookingId);
 
 	Payment addPayment(Payment payment);
 }
