+++ SimplyFly/src/main/java/com/hexaware/simplyfly/service/RouteService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.hexaware.simplyfly.entity.Route;
 
 public interface RouteService {
-	Route getRouteById(Long id);
+	Route getRouteById(int id);
 
 	List<Route> searchRoutes(String origin, String destination);
